refactor api end points
