+++ src/main/java/com/launchacademy/javaspringandreact/repositories/AdoptionApplicationRepository.java
@@ -1,9 +1,12 @@
 package com.launchacademy.javaspringandreact.repositories;
 
 import com.launchacademy.javaspringandreact.models.AdoptionApplication;
+import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 public interface AdoptionApplicationRepository extends
     PagingAndSortingRepository<AdoptionApplication, Integer> {
+
+  public List<AdoptionApplication> findAll();
 
 }
