+++ src/main/java/com/launchacademy/javaspringandreact/repositories/PetSurrenderApplicationRepository.java
@@ -1,9 +1,12 @@
 package com.launchacademy.javaspringandreact.repositories;
 
 import com.launchacademy.javaspringandreact.models.PetSurrenderApplication;
+import java.util.List;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
 public interface PetSurrenderApplicationRepository extends
     PagingAndSortingRepository<PetSurrenderApplication, Integer> {
+
+  public List<PetSurrenderApplication> findAll();
 
 }
